+++ Core/OfficeDevPnP.Core/Framework/Provisioning/Model/WebApiPermissionCollection.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OfficeDevPnP.Core.Framework.Provisioning.Model
+{
+    /// <summary>
+    /// Collection of WebApiPermission objects
+    /// </summary>
+    public partial class WebApiPermissionCollection : ProvisioningTemplateCollection<WebApiPermission>
+    {
+        /// <summary>
+        /// Constructor for WebApiPermissionCollection class
+        /// </summary>
+        /// <param name="parentTemplate">Parent provisioning template</param>
+        public WebApiPermissionCollection(ProvisioningTemplate parentTemplate) : base(parentTemplate)
+        {
+
+        }
+    }
+}