+++ CHANGELOG.md
@@ -23,6 +23,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Added support for SystemUpdate of taxonomy fields on list extension and item extension methods.
 - Added support for using the ClientWebPart client side web part to host "classic" SharePoint Add-ins on client side pages
 - Added support for new schema v.2018-05
+- Added support for Web API Permission in schema v.2018-05
+- Schema v.2018-05 is now the default schema for the PnP Provisioning Engine
 
 ### Changed